requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
Pillow>=9.0.0
pyinstaller>=5.0.0
